Allow cross unit operations with different types

+++ test/src/derived_unit_tests/velocity_tests.cc
@@ -134,9 +134,9 @@ TEST_CASE("GIVEN a value WHEN constructed with literal _m_p_h THEN result is a "
 TEST_CASE("GIVEN a length of type float WHEN divided by a value of type double "
           "THEN resulting type is a valocity of type float") {
 
-  SI::metre_t<float> a = 1.5_km;
-  SI::seconds_t<double> s = 1.0;
-  SI::metre_per_second_t<float> c = a / s;
+  SI::metre_t<long double> a = 1.5_km;
+  SI::seconds_t<float> s = 1.0;
+  SI::metre_per_second_t<long double> c = a / s;
   STATIC_REQUIRE(
-      std::is_same<decltype(c), SI::metre_per_second_t<float>>::value);
+      std::is_same<decltype(c), SI::metre_per_second_t<long double>>::value);
 }
